refactor(conversation_tab): extract OCR functionality into a separate module
