+++ basilisk/gui/conversation_tab.py
@@ -20,7 +20,7 @@ import logging
 import re
 import threading
 import time
-from multiprocessing import Process, Queue, Value
+from multiprocessing import Process
 from typing import TYPE_CHECKING, Any, Optional
 
 import wx
@@ -46,14 +46,13 @@ from basilisk.conversation import (
 	parse_supported_attachment_formats,
 )
 from basilisk.decorators import ensure_no_task_running
-from basilisk.process_helper import run_task
 from basilisk.provider_ai_model import ProviderAIModel
 from basilisk.provider_capability import ProviderCapability
 from basilisk.sound_manager import play_sound, stop_sound
 
 from .base_conversation import BaseConversation
 from .history_msg_text_ctrl import HistoryMsgTextCtrl
-from .progress_bar_dialog import ProgressBarDialog
+from .ocr_handler import OCRHandler
 from .read_only_message_dialog import ReadOnlyMessageDialog
 
 if TYPE_CHECKING:
@@ -171,6 +170,7 @@ class ConversationTab(wx.Panel, BaseConversation):
 		self.task = None
 		self.process: Optional[Process] = None
 		self._stop_completion = False
+		self.ocr_handler = OCRHandler(self)
 		self.init_ui()
 		self.init_data(profile)
 		self.adjust_advanced_mode_setting()
@@ -257,8 +257,10 @@ class ConversationTab(wx.Panel, BaseConversation):
 		self.attachments_list.SetColumnWidth(1, 100)
 		self.attachments_list.SetColumnWidth(2, 500)
 		sizer.Add(self.attachments_list, proportion=0, flag=wx.ALL | wx.EXPAND)
-		self.create_ocr_widget()
+
+		self.ocr_button = self.ocr_handler.create_ocr_widget(self)
 		sizer.Add(self.ocr_button, proportion=0, flag=wx.EXPAND)
+
 		label = self.create_model_widget()
 		sizer.Add(label, proportion=0, flag=wx.EXPAND)
 		sizer.Add(self.model_list, proportion=0, flag=wx.ALL | wx.EXPAND)
@@ -1151,6 +1153,11 @@ class ConversationTab(wx.Panel, BaseConversation):
 		}
 
 	def _check_attachments_valid(self) -> bool:
+		"""Check if all attachments are valid and supported by the current provider.
+
+		Returns:
+			True if all attachments are valid, False otherwise
+		"""
 		supported_attachment_formats = (
 			self.current_engine.supported_attachment_formats
 		)
@@ -1171,262 +1178,6 @@ class ConversationTab(wx.Panel, BaseConversation):
 				invalid_found = True
 		return not invalid_found
 
-	def create_ocr_widget(self):
-		"""Create and configure the OCR button widget."""
-		self.ocr_button = wx.Button(
-			self,
-			# Translators: This is a label for perform OCR button in the conversation tab
-			label=_("Perform OCR"),
-		)
-		self.ocr_button.Bind(wx.EVT_BUTTON, self.on_ocr)
-
-	def _handle_ocr_message(
-		self, message_type: str, data: Any, dialog: ProgressBarDialog
-	) -> None:
-		"""Handle a message from the OCR result queue.
-
-		Args:
-			message_type: The type of message received
-			data: The message data
-			dialog: The progress dialog
-		"""
-		try:
-			if message_type == "message":
-				self._handle_ocr_info_message(data, dialog)
-			elif message_type == "progress":
-				self._handle_ocr_progress_message(data, dialog)
-			elif message_type in ("result", "error"):
-				self._handle_ocr_completion_message(message_type, data, dialog)
-			else:
-				log.warning(
-					f"Unknown message type in result queue: {message_type}"
-				)
-		except Exception as e:
-			log.error(f"Error handling message: {str(e)}", exc_info=True)
-
-	def _handle_ocr_info_message(
-		self, data: Any, dialog: ProgressBarDialog
-	) -> None:
-		"""Handle an information message from the OCR process.
-
-		Args:
-			data: The message data
-			dialog: The progress dialog
-		"""
-		if isinstance(data, str) and data:
-			dialog.update_message(data)
-
-	def _handle_ocr_progress_message(
-		self, data: Any, dialog: ProgressBarDialog
-	) -> None:
-		"""Handle a progress update message from the OCR process.
-
-		Args:
-			data: The progress value
-			dialog: The progress dialog
-		"""
-		if isinstance(data, int):
-			dialog.update_progress_bar(data)
-
-	def _handle_ocr_completion_message(
-		self, message_type: str, data: Any, dialog: ProgressBarDialog
-	) -> None:
-		"""Handle a completion (result or error) message from the OCR process.
-
-		Args:
-			message_type: The type of message ("result" or "error")
-			data: The message data
-			dialog: The progress dialog
-		"""
-		if not dialog:
-			# Dialog might have been destroyed already
-			return
-
-		dialog.Destroy()
-		self.ocr_button.Enable()
-
-		# Handle based on message type
-		if message_type == "result":
-			self._display_ocr_result(data)
-		else:  # error case
-			wx.MessageBox(str(data), _("Error"), wx.OK | wx.ICON_ERROR)
-
-	def _display_ocr_result(self, data: Any) -> None:
-		"""Display the result of OCR processing.
-
-		Args:
-			data: The OCR result data
-		"""
-		if isinstance(data, list) and data:
-			msg = (
-				_("OCR completed successfully. Text extracted to:")
-				+ "\n"
-				+ "\n".join(data)
-			)
-			wx.MessageBox(msg, _("Result"), wx.OK | wx.ICON_INFORMATION)
-		elif isinstance(data, str) and data:
-			wx.MessageBox(data, _("Result"), wx.OK | wx.ICON_INFORMATION)
-		else:
-			log.warning(f"{data}")
-			wx.MessageBox(
-				_("OCR completed, but no text was extracted."),
-				_("Result"),
-				wx.OK | wx.ICON_INFORMATION,
-			)
-
-	def _process_ocr_queue(
-		self, result_queue: Queue, dialog: ProgressBarDialog
-	) -> None:
-		"""Process all pending messages in the OCR result queue.
-
-		Args:
-			result_queue: The queue containing OCR process results
-			dialog: The progress dialog to update
-		"""
-		try:
-			while not result_queue.empty():
-				message_type, data = result_queue.get(block=False)
-				self._handle_ocr_message(message_type, data, dialog)
-		except Exception as e:
-			log.error(
-				f"Error processing queue messages: {str(e)}", exc_info=True
-			)
-
-	def _cleanup_ocr_process(self, dialog: ProgressBarDialog) -> None:
-		"""Clean up OCR process resources when task completes or is canceled.
-
-		Args:
-			dialog: The progress dialog to manage
-		"""
-		# Clean up resources
-		if (
-			self.process
-			and dialog.cancel_flag.value
-			and self.process.is_alive()
-		):
-			log.debug("Terminating OCR process due to user cancellation")
-			try:
-				self.process.terminate()
-				self.process.join(timeout=1.0)  # Wait for process to terminate
-				if self.process.is_alive():
-					log.warning("Process did not terminate, killing it")
-					self.process.kill()
-			except Exception as e:
-				log.error(f"Error terminating process: {str(e)}", exc_info=True)
-
-		try:
-			if dialog and dialog.IsShown():
-				dialog.Destroy()
-		except Exception as e:
-			log.error(f"Error destroying dialog: {str(e)}", exc_info=True)
-
-		self.ocr_button.Enable()
-		self.process = None
-
-	def check_task_progress(
-		self, dialog: ProgressBarDialog, result_queue: Queue, cancel_flag
-	):
-		"""Check the progress of the OCR task.
-
-		Args:
-			dialog: The progress dialog
-			result_queue: The queue to store the task result
-			cancel_flag: The flag to indicate if the task should be cancelled
-		"""
-		# Process all pending messages in the queue
-		self._process_ocr_queue(result_queue, dialog)
-
-		# Check if process is still running
-		if (
-			not self.process
-			or not self.process.is_alive()
-			or dialog.cancel_flag.value
-		):
-			self._cleanup_ocr_process(dialog)
-		else:
-			# Continue checking progress
-			wx.CallLater(
-				CHECK_TASK_DELAY,
-				self.check_task_progress,
-				dialog,
-				result_queue,
-				cancel_flag,
-			)
-
-	def on_ocr(self, event: wx.CommandEvent):
-		"""Handle the OCR button click event.
-
-		Args:
-			event: The button click event
-		"""
-		engine: BaseEngine = self.current_engine
-		if ProviderCapability.OCR not in engine.capabilities:
-			wx.MessageBox(
-				# Translators: This message is displayed when the current provider does not support OCR.
-				_("The selected provider does not support OCR."),
-				_("Error"),
-				wx.OK | wx.ICON_ERROR,
-			)
-			return
-
-		if not self.attachment_files:
-			wx.MessageBox(
-				# Translators: This message is displayed when there are no attachments to perform OCR on.
-				_("No attachments to perform OCR on."),
-				_("Error"),
-				wx.OK | wx.ICON_ERROR,
-			)
-			return
-		if not self._check_attachments_valid():
-			return
-		client = engine.client
-		if not client:
-			wx.MessageBox(
-				# Translators: This message is displayed when the current provider does not have a client.
-				_("The selected provider does not have a client."),
-				_("Error"),
-				wx.OK | wx.ICON_ERROR,
-			)
-			return
-		self.ocr_button.Disable()
-
-		cancel_flag = Value('i', 0)
-		result_queue = Queue()
-
-		progress_bar_dialog = ProgressBarDialog(
-			self,
-			title=_("Performing OCR..."),
-			message=_("Performing OCR on attachments..."),
-			cancel_flag=cancel_flag,
-		)
-		progress_bar_dialog.Show()
-
-		kwargs = {
-			"api_key": self.current_account.api_key.get_secret_value(),
-			"base_url": self.current_account.custom_base_url
-			or self.current_account.provider.base_url,
-			"attachments": self.attachment_files,
-		}
-
-		self.process = Process(
-			target=run_task,
-			args=(engine.handle_ocr, result_queue, cancel_flag),
-			kwargs=kwargs,
-		)
-
-		self.process.daemon = True  # Ensure process terminates when parent does
-		self.process.start()
-		log.debug(f"OCR process started: {self.process.pid}")
-
-		wx.CallLater(
-			CHECK_TASK_DELAY,
-			self.check_task_progress,
-			progress_bar_dialog,
-			result_queue,
-			cancel_flag,
-		)
-
-	@ensure_no_task_running
 	def on_submit(self, event: wx.CommandEvent):
 		"""Handle the submission of a new message block for completion.
